moved input fields to an internal folder
