+++ aqueductcore/frontend/src/components/atoms/TextAreaField/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Grid, TextareaAutosize, TextareaAutosizeProps, styled } from "@mui/material";
 
-import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents"
+import { FieldDescription, FieldTitle, FieldType } from "components/atoms/sharedStyledComponents/pluginInputFields"
 import { PluginFieldBase } from "types/globalTypes";
 
 const TextAreaInput = styled(TextareaAutosize)`
